+++ src/main/java/dev/strubbelkopp/simple_quiver/client/IMouse.java
@@ -1,0 +1,5 @@
+package dev.strubbelkopp.simple_quiver.client;
+
+public interface IMouse {
+    void quiver$setArrowCycleState(boolean state);
+}